Move Header and NewHeader into headerv1.go
This file will move into block/v1 directory.
